Added gallery builders layout
