+++ apps/scoutgame/lib/builders/getPaginatedBuildersAction.ts
@@ -5,14 +5,13 @@ import type { BuilderInfo } from '@packages/scoutgame/builders/interfaces';
 import { currentSeason, getCurrentWeek } from '@packages/scoutgame/dates';
 import * as yup from 'yup';
 
-import type { CompositeCursor } from 'lib/builders/getSortedBuilders';
-import { getSortedBuilders } from 'lib/builders/getSortedBuilders';
+import type { CompositeCursor } from 'lib/builders/getPaginatedBuilders';
+import { getPaginatedBuilders } from 'lib/builders/getPaginatedBuilders';
 
-export const getSortedBuildersAction = actionClient
-  .metadata({ actionName: 'get_sorted_builders' })
+export const getPaginatedBuildersAction = actionClient
+  .metadata({ actionName: 'get_paginated_builders' })
   .schema(
     yup.object({
-      sort: yup.string().oneOf(['hot', 'new', 'top']).required(),
       cursor: yup
         .object({
           userId: yup.string().required(),
@@ -22,9 +21,8 @@ export const getSortedBuildersAction = actionClient
     })
   )
   .action<{ builders: BuilderInfo[]; nextCursor: CompositeCursor | null }>(async ({ parsedInput }) => {
-    const { sort, cursor } = parsedInput;
-    const { builders, nextCursor } = await getSortedBuilders({
-      sort,
+    const { cursor } = parsedInput;
+    const { builders, nextCursor } = await getPaginatedBuilders({
       limit: 30, // 6 rows per page
       week: getCurrentWeek(),
       season: currentSeason,
